refactor globals
