+++ src/llms/groq.ts
@@ -2,7 +2,7 @@
 
 import { writeFile } from 'node:fs/promises'
 import { env } from 'node:process'
-import { GROQ_MODELS } from '../utils/globals'
+import { GROQ_MODELS } from '../utils/llm-models'
 import { err, logAPIResults } from '../utils/logging'
 import type { LLMFunction, GroqModelType, GroqChatCompletionResponse } from '../types/llms'
